some fixes. move files to notifications folder
