+++ api/lib/block-listeners/polkadot-node-subscription.js
@@ -9,7 +9,7 @@ const {
   lunieMessageTypes: { SEND }
 } = require('../message-types.js')
 // const { eventTypes } = require('../notifications/pushNotifications-format')
-const { eventTypes, resourceTypes } = require('../notifications-types')
+const { eventTypes, resourceTypes } = require('../notifications/notifications-types')
 const Sentry = require('@sentry/node')
 const database = require('../database')
 const config = require('../../config.js')
